requirements.txt
fastapi
uvicorn
python-multipart
opencv-python
onnxruntime
numpy
pillow
torch
torchvision